third_party/WebKit/LayoutTests/crypto/subtle/clone-ecKey-public-expected.txt

Tests structured cloning of EC public keys

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "ECDSA"
PASS importedKey.algorithm.namedCurve is "P-256"
PASS importedKey.usages.join(',') is ""
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "ECDSA"
PASS clonedKey.algorithm.namedCurve is "P-256"
PASS clonedKey.usages.join(',') is ""
Serialized key bytes: 3f004b050e0101015b3059301306072a8648ce3d020106082a8648ce3d030107034200049cb0cf69303dafc761d4e4687b4ecf039e6d34ab964af80810d8d558a4a8d6f72d51233a1788920a86ee08a1962c79efa317fb7879e297dad2146db995fa1c78
PASS: Cloned key exported data should be [3059301306072a8648ce3d020106082a8648ce3d030107034200049cb0cf69303dafc761d4e4687b4ecf039e6d34ab964af80810d8d558a4a8d6f72d51233a1788920a86ee08a1962c79efa317fb7879e297dad2146db995fa1c78] and was

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "ECDSA"
PASS importedKey.algorithm.namedCurve is "P-384"
PASS importedKey.usages.join(',') is ""
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "ECDSA"
PASS clonedKey.algorithm.namedCurve is "P-384"
PASS clonedKey.usages.join(',') is ""
Serialized key bytes: 3f004b050e010201783076301006072a8648ce3d020106052b81040022036200040874a2e0b8ff448f0e54321e27f4f1e64d064cdeb7d26f458c32e930120f4e57dc85c2693f977eed4a8ecc8db981b4d91f69446df4f4c6f5de19003f45f891d0ebcd2fffdb5c81c040e8d6994c43c7feedb98a4a31edfb35e89a30013c3b926700
PASS: Cloned key exported data should be [3076301006072a8648ce3d020106052b81040022036200040874a2e0b8ff448f0e54321e27f4f1e64d064cdeb7d26f458c32e930120f4e57dc85c2693f977eed4a8ecc8db981b4d91f69446df4f4c6f5de19003f45f891d0ebcd2fffdb5c81c040e8d6994c43c7feedb98a4a31edfb35e89a30013c3b9267] and was

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "ECDSA"
PASS importedKey.algorithm.namedCurve is "P-521"
PASS importedKey.usages.join(',') is ""
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "ECDSA"
PASS clonedKey.algorithm.namedCurve is "P-521"
PASS clonedKey.usages.join(',') is ""
Serialized key bytes: 3f004b050e0103019e0130819b301006072a8648ce3d020106052b81040023038186000400f50a08703250c15f043c8c46e99783435245cf98f4f2694b0e2f8d029a514dd6f0b086d4ed892000cd5590107aae69c4c0a7a95f7cf74e5770a07d5db55bce4ab400f2c770bab8b9be4cdb6ecd3dc26c698da0d2599cebf3d904f7f9ca3a55e64731810d73cd317264e50baba4bc2860857e16d6cbb79501bc9e3a32bd172ea8a71dee
PASS: Cloned key exported data should be [30819b301006072a8648ce3d020106052b81040023038186000400f50a08703250c15f043c8c46e99783435245cf98f4f2694b0e2f8d029a514dd6f0b086d4ed892000cd5590107aae69c4c0a7a95f7cf74e5770a07d5db55bce4ab400f2c770bab8b9be4cdb6ecd3dc26c698da0d2599cebf3d904f7f9ca3a55e64731810d73cd317264e50baba4bc2860857e16d6cbb79501bc9e3a32bd172ea8a71dee] and was

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "ECDSA"
PASS importedKey.algorithm.namedCurve is "P-256"
PASS importedKey.usages.join(',') is "verify"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "ECDSA"
PASS clonedKey.algorithm.namedCurve is "P-256"
PASS clonedKey.usages.join(',') is "verify"
Serialized key bytes: 3f004b050e0101115b3059301306072a8648ce3d020106082a8648ce3d030107034200049cb0cf69303dafc761d4e4687b4ecf039e6d34ab964af80810d8d558a4a8d6f72d51233a1788920a86ee08a1962c79efa317fb7879e297dad2146db995fa1c78
PASS: Cloned key exported data should be [3059301306072a8648ce3d020106082a8648ce3d030107034200049cb0cf69303dafc761d4e4687b4ecf039e6d34ab964af80810d8d558a4a8d6f72d51233a1788920a86ee08a1962c79efa317fb7879e297dad2146db995fa1c78] and was

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "ECDSA"
PASS importedKey.algorithm.namedCurve is "P-384"
PASS importedKey.usages.join(',') is "verify"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "ECDSA"
PASS clonedKey.algorithm.namedCurve is "P-384"
PASS clonedKey.usages.join(',') is "verify"
Serialized key bytes: 3f004b050e010211783076301006072a8648ce3d020106052b81040022036200040874a2e0b8ff448f0e54321e27f4f1e64d064cdeb7d26f458c32e930120f4e57dc85c2693f977eed4a8ecc8db981b4d91f69446df4f4c6f5de19003f45f891d0ebcd2fffdb5c81c040e8d6994c43c7feedb98a4a31edfb35e89a30013c3b926700
PASS: Cloned key exported data should be [3076301006072a8648ce3d020106052b81040022036200040874a2e0b8ff448f0e54321e27f4f1e64d064cdeb7d26f458c32e930120f4e57dc85c2693f977eed4a8ecc8db981b4d91f69446df4f4c6f5de19003f45f891d0ebcd2fffdb5c81c040e8d6994c43c7feedb98a4a31edfb35e89a30013c3b9267] and was

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "ECDSA"
PASS importedKey.algorithm.namedCurve is "P-521"
PASS importedKey.usages.join(',') is "verify"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "ECDSA"
PASS clonedKey.algorithm.namedCurve is "P-521"
PASS clonedKey.usages.join(',') is "verify"
Serialized key bytes: 3f004b050e0103119e0130819b301006072a8648ce3d020106052b81040023038186000400f50a08703250c15f043c8c46e99783435245cf98f4f2694b0e2f8d029a514dd6f0b086d4ed892000cd5590107aae69c4c0a7a95f7cf74e5770a07d5db55bce4ab400f2c770bab8b9be4cdb6ecd3dc26c698da0d2599cebf3d904f7f9ca3a55e64731810d73cd317264e50baba4bc2860857e16d6cbb79501bc9e3a32bd172ea8a71dee
PASS: Cloned key exported data should be [30819b301006072a8648ce3d020106052b81040023038186000400f50a08703250c15f043c8c46e99783435245cf98f4f2694b0e2f8d029a514dd6f0b086d4ed892000cd5590107aae69c4c0a7a95f7cf74e5770a07d5db55bce4ab400f2c770bab8b9be4cdb6ecd3dc26c698da0d2599cebf3d904f7f9ca3a55e64731810d73cd317264e50baba4bc2860857e16d6cbb79501bc9e3a32bd172ea8a71dee] and was

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "ECDSA"
PASS importedKey.algorithm.namedCurve is "P-256"
PASS importedKey.usages.join(',') is ""
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "ECDSA"
PASS clonedKey.algorithm.namedCurve is "P-256"
PASS clonedKey.usages.join(',') is ""
Serialized key bytes: 3f004b050e0101005b3059301306072a8648ce3d020106082a8648ce3d030107034200049cb0cf69303dafc761d4e4687b4ecf039e6d34ab964af80810d8d558a4a8d6f72d51233a1788920a86ee08a1962c79efa317fb7879e297dad2146db995fa1c78

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "ECDSA"
PASS importedKey.algorithm.namedCurve is "P-384"
PASS importedKey.usages.join(',') is ""
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "ECDSA"
PASS clonedKey.algorithm.namedCurve is "P-384"
PASS clonedKey.usages.join(',') is ""
Serialized key bytes: 3f004b050e010200783076301006072a8648ce3d020106052b81040022036200040874a2e0b8ff448f0e54321e27f4f1e64d064cdeb7d26f458c32e930120f4e57dc85c2693f977eed4a8ecc8db981b4d91f69446df4f4c6f5de19003f45f891d0ebcd2fffdb5c81c040e8d6994c43c7feedb98a4a31edfb35e89a30013c3b926700

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "ECDSA"
PASS importedKey.algorithm.namedCurve is "P-521"
PASS importedKey.usages.join(',') is ""
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "ECDSA"
PASS clonedKey.algorithm.namedCurve is "P-521"
PASS clonedKey.usages.join(',') is ""
Serialized key bytes: 3f004b050e0103009e0130819b301006072a8648ce3d020106052b81040023038186000400f50a08703250c15f043c8c46e99783435245cf98f4f2694b0e2f8d029a514dd6f0b086d4ed892000cd5590107aae69c4c0a7a95f7cf74e5770a07d5db55bce4ab400f2c770bab8b9be4cdb6ecd3dc26c698da0d2599cebf3d904f7f9ca3a55e64731810d73cd317264e50baba4bc2860857e16d6cbb79501bc9e3a32bd172ea8a71dee

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "ECDSA"
PASS importedKey.algorithm.namedCurve is "P-256"
PASS importedKey.usages.join(',') is "verify"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "ECDSA"
PASS clonedKey.algorithm.namedCurve is "P-256"
PASS clonedKey.usages.join(',') is "verify"
Serialized key bytes: 3f004b050e0101105b3059301306072a8648ce3d020106082a8648ce3d030107034200049cb0cf69303dafc761d4e4687b4ecf039e6d34ab964af80810d8d558a4a8d6f72d51233a1788920a86ee08a1962c79efa317fb7879e297dad2146db995fa1c78

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "ECDSA"
PASS importedKey.algorithm.namedCurve is "P-384"
PASS importedKey.usages.join(',') is "verify"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "ECDSA"
PASS clonedKey.algorithm.namedCurve is "P-384"
PASS clonedKey.usages.join(',') is "verify"
Serialized key bytes: 3f004b050e010210783076301006072a8648ce3d020106052b81040022036200040874a2e0b8ff448f0e54321e27f4f1e64d064cdeb7d26f458c32e930120f4e57dc85c2693f977eed4a8ecc8db981b4d91f69446df4f4c6f5de19003f45f891d0ebcd2fffdb5c81c040e8d6994c43c7feedb98a4a31edfb35e89a30013c3b926700

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "public"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "ECDSA"
PASS importedKey.algorithm.namedCurve is "P-521"
PASS importedKey.usages.join(',') is "verify"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "public"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "ECDSA"
PASS clonedKey.algorithm.namedCurve is "P-521"
PASS clonedKey.usages.join(',') is "verify"
Serialized key bytes: 3f004b050e0103109e0130819b301006072a8648ce3d020106052b81040023038186000400f50a08703250c15f043c8c46e99783435245cf98f4f2694b0e2f8d029a514dd6f0b086d4ed892000cd5590107aae69c4c0a7a95f7cf74e5770a07d5db55bce4ab400f2c770bab8b9be4cdb6ecd3dc26c698da0d2599cebf3d904f7f9ca3a55e64731810d73cd317264e50baba4bc2860857e16d6cbb79501bc9e3a32bd172ea8a71dee

PASS successfullyParsed is true

TEST COMPLETE
